% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgVertDep.R
\name{avgVertDep}
\alias{avgVertDep}
\title{Calculation of the average vertex depth for rooted trees}
\usage{
avgVertDep(tree)
}
\arguments{
\item{tree}{A rooted tree in phylo format.}
}
\value{
\code{avgVertDep} returns the average vertex depth of the given tree.
}
\description{
This function calculates the average vertex depth \eqn{AVD(T)} for a given rooted
tree \eqn{T}. The tree must not necessarily be binary. \eqn{AVD(T)} is defined as
\deqn{AVD(T)=\frac{1}{|V(T)|}\cdot\sum_{x\in V(T)} \delta(x)}{AVD(T)=1/|V(T)|*\sum_{x\in V(T)} \delta(x)} in
which \eqn{V(T)} denotes the set of vertices of \eqn{T}, and \eqn{\delta(x)}{depth(x)}
denotes the depth of the vertex \eqn{x}. The average vertex depth is an
imbalance index.\cr\cr
For \eqn{n=1} the function returns \eqn{AVD(T)=0} and a warning.
}
\examples{
tree <- ape::read.tree(text="((((,),),(,)),(((,,),),(,)));")
avgVertDep(tree)

}
\references{
A. Herrada et al. Scaling properties of protein family phylogenies. BMC Evolutionary Biology, 11(1), June 2011. doi: 10.1186/1471-2148-11-155.
}
\author{
Luise Kuehn
}
