% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxWidth.R
\name{maxWidth}
\alias{maxWidth}
\title{Calculation of the maximum width of the tree}
\usage{
maxWidth(tree)
}
\arguments{
\item{tree}{A rooted tree in phylo format.}
}
\value{
\code{maxWidth} returns the maximum width of a tree.
}
\description{
This function calculates the maximum width \eqn{maxWidth(T)} for a
given rooted tree \eqn{T}. The tree must not necessarily be binary.
\eqn{maxWidth(T)} is defined as \deqn{maxWidth(T)=\max_{i=0,...,h(T)} w(i)}{maxWidth(T)=max_{i=0,...,h(T)} w(i)}
in which \eqn{h(T)} denotes the height of the tree \eqn{T} and \eqn{w(i)} denotes
the number of vertices in \eqn{T} that have depth \eqn{i}. The maximum width
is a balance index.
}
\examples{
tree <- ape::read.tree(text="((((,),),(,)),(((,),),(,)));")
maxWidth(tree)
tree <- ape::read.tree(text="((,),((((,),),),(,)));")
maxWidth(tree)

}
\references{
C. Colijn and J. Gardy.  Phylogenetic tree shapes resolve disease transmission patterns. Evolution, Medicine, and Public Health, 2014(1):96-108, 2014. ISSN 2050-6201. doi: 10.1093/emph/eou018.
}
\author{
Sophie Kersting
}
