% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mWovermD.R
\name{mWovermD}
\alias{mWovermD}
\title{Calculation of the maximum width over maximum depth of the tree}
\usage{
mWovermD(tree)
}
\arguments{
\item{tree}{A rooted tree in phylo format.}
}
\value{
\code{mWovermD} returns the maximum width over maximum depth of a tree.
}
\description{
This function calculates the maximum width over maximum depth \eqn{mWovermD(T)} for a
given rooted tree \eqn{T}. The tree must not necessarily be binary. For \eqn{n>1},
\eqn{mWovermD(T)} is defined as \deqn{mWovermD(T)=maxWidth(T) / h(T)}
in which \eqn{h(T)} denotes the height of the tree \eqn{T}, which is the same as the 
maximum depth of any leaf in the tree, and \eqn{maxWidth(T)} denotes
the maximum width of the tree \eqn{T}. The maximum width over maximum depth
is a balance index.
}
\examples{
tree <- ape::read.tree(text="((((,),),(,)),(((,),),(,)));")
mWovermD(tree)
tree <- ape::read.tree(text="((,),((((,),),),(,)));")
mWovermD(tree)

}
\author{
Luise Kuehn
}
