% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totPathLen.R
\name{totPathLen}
\alias{totPathLen}
\title{Calculation of the total path length for rooted trees}
\usage{
totPathLen(tree)
}
\arguments{
\item{tree}{A rooted tree in phylo format.}
}
\value{
\code{totPathLen} returns the total path length of the given tree.
}
\description{
This function calculates the total path length \eqn{TPL(T)} for a given rooted
tree \eqn{T}. The tree must not necessarily be binary. \eqn{TPL(T)} is defined as
\deqn{TPL(T)=\sum_{x\in V(T)} \delta(x)} in
which \eqn{V(T)} denotes the set of vertices of \eqn{T}, and \eqn{\delta(x)}{depth(x)}
denotes the depth of the vertex \eqn{x}. The total path length is an
imbalance index.\cr\cr
For \eqn{n=1} the function returns \eqn{TPL(T)=0} and a warning.
}
\examples{
tree <- ape::read.tree(text="((((,),),(,)),(((,,),),(,)));")
totPathLen(tree)

}
\references{
TODO
}
\author{
Luise Kuehn
}
