% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedEth-data.R
\docType{data}
\name{wedEth}
\alias{wedEth}
\title{Wedderburn Etherington numbers (from OEIS)}
\format{
numerical vector
}
\source{
OEIS Sequence A001190 available at https://oeis.org/A001190
}
\usage{
data(wedEth)
}
\description{
Contains a vector of Wedderburn Etherington numbers for \eqn{n=1} to \eqn{n=2545}.
}
\examples{
data(wedEth)
wedEth[5]
}
\keyword{datasets}
